fixede time
